Matt| All sorting should work correctly now. also shanged Pets weight from int to double. Only thing left is fixing printing spacing and do the sorting

+++ Pet.h
@@ -9,7 +9,7 @@ class Pet{
 public:
     string name;
     string type;
-    unsigned int weight;
+    double weight;
     double price;
 public:
     // Default Constructor
